backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.10.0
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.20
pymupdf==1.24.14
python-docx==1.1.2
